requirements.txt
# FastAPI et serveur
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Twilio pour WhatsApp
twilio==8.10.0

# LangChain et LLM
langchain>=0.1.0
langchain-core>=0.1.7
langchain-mistralai>=0.0.2
langchain-community>=0.0.10

# ChromaDB pour RAG
chromadb>=0.5.0

# Data science et ML
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Utilitaires
python-dotenv==1.0.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Logging et monitoring
python-json-logger==2.0.7

# Traitement de documents
python-docx>=1.1.0
pypdf>=3.17.0

# Visualisation (pour les modèles DIORES)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
missingno>=0.5.0

